change monitor script and fix nb of test accordingly to the new delay

+++ scripts/ac-debug/ac-xivo-monitor-script.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-# EA - 20130610
+# EA - 20131204
 # Needs :
 #  - dstat
 #  - iotop
@@ -8,26 +8,37 @@
 #  Launch both utilities in logs in /var/local/
 #  plus it removes log older than 30 days
 # Usage :
-#  Can be used in a crontab like :
+#  Can be used in a crontab like (for a 24 hour log with a line each 15s) :
 ## # Cron which launches monitor utility
 ## MAILTO=""
 ## 
 ## ## Launch two monitor utilities dstat and iotop
-## 02 0 * * * root /usr/local/sbin/ac-xivo-monitor-script.sh 2870
+## 02 0 * * * root /usr/local/sbin/ac-xivo-monitor-script.sh 24 15
 
 
 dstat_fct() {
-    /usr/bin/dstat -tcdgilmnf --output /var/local/`date +%Y%m%d`-`hostname`-dstat.log 15 $1
+    /usr/bin/dstat -tcdgilmnf --output /var/local/`date +%Y%m%d`-`hostname`-dstat.log $1 $2
 }
 
 iotop_fct() {
-    /usr/bin/iotop -b -o -t -d 15 -n $1 > /var/local/`date +%Y%m%d`-`hostname`-iotop.log
+    /usr/bin/iotop -b -o -t -d $1 -n $2 > /var/local/`date +%Y%m%d`-`hostname`-iotop.log
 }
 
-COUNT_NB=$1
+if [ $# -ne 2 ]; then
+    echo "Usage: $0 DURATION DELAY"
+    echo "Where :"
+    echo "  - DURATION : how many HOURS the test will be run,"
+    echo "  - DELAY : how many SECONDS between each line of log"
+    exit 1
+fi
 
-dstat_fct $COUNT_NB &
-iotop_fct $COUNT_NB &
+DURATION=$1
+DELAY=$2
+
+COUNT_NB=$((${DURATION} * 60 * 60 / ${DELAY}))
+
+dstat_fct $DELAY $COUNT_NB &
+iotop_fct $DELAY $COUNT_NB &
 
 find /var/local/*{dstat,iotop}.log -mtime +30 -exec rm -f {} \;
